feat : anonymous swap in/out

+++ include/vm/anon.h
@@ -6,7 +6,7 @@ enum vm_type;
 
 struct anon_page {
     /* Project 3 : VM */
-    int swap_slot_index;
+    size_t swap_slot_index;
 };
 
 void vm_anon_init (void);
